documents_ocr_server/documents_ocr_server/requirements.txt
cffi==1.15.0
chardet==4.0.0
coloredlogs==15.0.1
cryptography==36.0.1
humanfriendly==10.0
img2pdf==0.4.3
importlib-resources==5.4.0
lxml==4.7.1
numpy==1.22.1
ocrmypdf==13.4.2
opencv-python-headless==4.5.5.62
packaging==21.3
pdf2image==1.16.0
pdfminer.six==20211012
pdfplumber==0.6.0
pikepdf==4.3.1
Pillow==9.0.0
pkg_resources==0.0.0
pluggy==1.0.0
pycparser==2.21
pyparsing==3.0.6
PyPDF2==1.26.0
pytesseract==0.3.8
reportlab==3.6.5
tqdm==4.62.3
Wand==0.6.7
zipp==3.7.0
